langchain prompts/requirements.txt
# LangChain Core
langchain
langchain-core

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai


# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# UI
streamlit